# Issue 1225 - Poor performance of estimate u in a link_only job