--- a/tests/test_join_type_for_estimate_u_and_predict_are_efficient.py
+++ b/tests/test_join_type_for_estimate_u_and_predict_are_efficient.py
@@ -72,7 +72,7 @@ data_three = [
     {
         "unique_id": "b1",
         "first_name": "Juli",
-        "surname": "Taylor",
+        "surname": "Taylr",
         "dob": "2015-07-31",
         "city": "London",
         "email": "hannah8@powers.com",