### Type of PR

- [ ] BUG
- [x] FEAT
- [ ] MAINT
- [ ] DOC

### Is your Pull Request linked to an existing Issue or Pull Request?

Closes #1225 - see the issue for a detailed description of the underlying cause of the problem

Built on top of #1059 so let's get that merged first, then we can point this at `master`

### Give a brief description for the solution you have provided

The challenge is that for a link only, it's much more efficient to run sql like:
```
select ...
from df_one
inner join
df_two
[where conditions]
```

rather than concatenating the two and running:
```
select ...
from dfs_concat as l
inner join
dfs_concat as r
where l.source_dataset != r.source_dataset
[other where conditions]
```

The later  is inefficient because intra-dataset comparisons are generated and then rejected, rather than never being generated.

The efficient solution is fine for the n=2 case, but in n=3 you'd need (pseudocode):
```
df_one inner join df_two
union all
df_one inner join df_three
union all
df_two inner join df_three
```

This statement will quickly become unmanageable as n grows

Options:

1.  Could try to implement a left join approach for the n=2 case (which is most common).  This is relatively straightforward in the case of 2 input datasets, but results in n choose 2 joins in the more general case of n datasets.

Pros:

- Consistent with the approach I currently use for blocking. 

Cons:
- Doesn't solve the n>2 case.

2. In the case if unbalanced datasets, (we could test when the ratio of target rows to total comparisons is poor), change link_type to 'link and dedupe' for the purpose of estimating u

Pros:
- Generalise to n>2

Cons:
- Questionable estimates of u


### Conclusion

For the moment I've gone with option 1 since it preserves the quality of estimates.

Note that the code still works for n>2, it just uses the less inefficient `where l.source_dataset != r.source_dataset` which is  n't too bad in most cases, but gets extremely bad in the case of very unbalanced datsets.


### PR Checklist

- [ ] Added tests (if appropriate)