@RobinL 2023-06-24
Remember that once https://github.com/moj-analytical-services/splink/issues/1061 is merged, it will be possible to write tests of this using the sql that has been generated