@ThomasHepworth 2023-06-26
### Rough Test Results

Some very rough performance metrics for this change (done locally and not w/ pytest):
| df_l size | df_r size | old performance | new performance |
|-----------|-----------|-----------------|-----------------|
| 1000      | 512,000   | 1220 seconds    | 55 seconds     |
| 256,000   | 256,000   | 70 seconds      | 52 seconds      |

For the heavily unbalanced job, this new method appears to be about 24x faster.

It also appears to have an indirect impact on the performance of standard u training jobs, being approx 34% faster in the test below.

<hr>

### Test Code

```python
import time
from splink.duckdb.duckdb_linker import DuckDBLinker
import pandas as pd

from tests.basic_settings import get_settings_dict

df_l = pd.read_csv("./tests/datasets/fake_1000_from_splink_demos.csv")
df_r = df_l.copy()

for _ in range(9):
    df_r = pd.concat([df_r, df_r])

settings = get_settings_dict()
settings["link_type"] = "link_only"

# Equally sized link
# linker = DuckDBLinker(
#     [df_r, df_r],
#     settings,
# )

# Heavily weighted df on right
linker = DuckDBLinker(
    [df_l, df_r],
    settings,
)

start = time.time()
linker.estimate_u_using_random_sampling(target_rows=1e8)
display(time.time()-start)
```